chore(routes): move beta routes to separate dir
