+++ src/openai_responses/_routes/beta/messages.py
@@ -9,15 +9,15 @@ from openai.types.beta.threads.message import Message
 from openai.types.beta.threads.message_deleted import MessageDeleted
 from openai.types.beta.threads.message_update_params import MessageUpdateParams
 
-from ._base import StatefulRoute
+from .._base import StatefulRoute
 
-from ..stores import StateStore
-from .._types.partials.sync_cursor_page import PartialSyncCursorPage
-from .._types.partials.messages import PartialMessage, PartialMessageDeleted
+from ...stores import StateStore
+from ..._types.partials.sync_cursor_page import PartialSyncCursorPage
+from ..._types.partials.messages import PartialMessage, PartialMessageDeleted
 
-from .._utils.faker import faker
-from .._utils.serde import json_loads, model_dict, model_parse
-from .._utils.time import utcnow_unix_timestamp_s
+from ..._utils.faker import faker
+from ..._utils.serde import json_loads, model_dict, model_parse
+from ..._utils.time import utcnow_unix_timestamp_s
 
 __all__ = [
     "MessageCreateRoute",
