+++ src/openai_responses/_routes/beta/threads.py
@@ -10,20 +10,20 @@ from openai.types.beta.thread_create_params import ThreadCreateParams
 from openai.types.beta.thread_update_params import ThreadUpdateParams
 from openai.types.beta.thread_deleted import ThreadDeleted
 
-from ._base import StatefulRoute
+from .._base import StatefulRoute
 
-from ..helpers.builders.messages import message_from_create_request
-from ..helpers.builders.vector_stores import vector_store_from_create_request
-from ..helpers.builders.vector_store_files import vector_store_file_from_create_request
-from ..helpers.mergers.threads import merge_thread_with_partial
+from ...helpers.builders.messages import message_from_create_request
+from ...helpers.builders.vector_stores import vector_store_from_create_request
+from ...helpers.builders.vector_store_files import vector_store_file_from_create_request
+from ...helpers.mergers.threads import merge_thread_with_partial
 
-from ..stores import StateStore
-from .._types.partials.deleted import PartialResourceDeleted
-from .._types.partials.threads import PartialThread
+from ...stores import StateStore
+from ..._types.partials.deleted import PartialResourceDeleted
+from ..._types.partials.threads import PartialThread
 
-from .._utils.faker import faker
-from .._utils.serde import json_loads, model_dict, model_parse
-from .._utils.time import utcnow_unix_timestamp_s
+from ..._utils.faker import faker
+from ..._utils.serde import json_loads, model_dict, model_parse
+from ..._utils.time import utcnow_unix_timestamp_s
 
 
 __all__ = [
