+++ src/openai_responses/_routes/beta/vector_store_file_batches.py
@@ -13,18 +13,18 @@ from openai.types.beta.vector_stores.file_batch_create_params import (
 )
 
 
-from ._base import StatefulRoute
+from .._base import StatefulRoute
 
-from ..helpers.builders.vector_store_files import vector_store_file_from_create_request
+from ...helpers.builders.vector_store_files import vector_store_file_from_create_request
 
-from ..stores import StateStore
-from .._types.partials.sync_cursor_page import PartialSyncCursorPage
-from .._types.partials.vector_store_files import PartialVectorStoreFile
-from .._types.partials.vector_store_file_batches import PartialVectorStoreFileBatch
+from ...stores import StateStore
+from ..._types.partials.sync_cursor_page import PartialSyncCursorPage
+from ..._types.partials.vector_store_files import PartialVectorStoreFile
+from ..._types.partials.vector_store_file_batches import PartialVectorStoreFileBatch
 
-from .._utils.faker import faker
-from .._utils.serde import json_loads, model_dict, model_parse
-from .._utils.time import utcnow_unix_timestamp_s
+from ..._utils.faker import faker
+from ..._utils.serde import json_loads, model_dict, model_parse
+from ..._utils.time import utcnow_unix_timestamp_s
 
 __all__ = [
     "VectorStoreFileBatchCreateRoute",
