+++ src/openai_responses/helpers/builders/runs.py
@@ -5,7 +5,7 @@ import httpx
 from openai.types.beta.threads.run import Run
 
 from ._base import _generic_builder
-from ..._routes.runs import RunCreateRoute
+from ..._routes.beta.runs import RunCreateRoute
 from ..._types.partials.runs import PartialRun
 
 __all__ = ["run_from_create_request"]
