+++ src/openai_responses/helpers/builders/threads.py
@@ -5,7 +5,7 @@ import httpx
 from openai.types.beta.thread import Thread
 
 from ._base import _generic_builder
-from ..._routes.threads import ThreadCreateRoute
+from ..._routes.beta.threads import ThreadCreateRoute
 from ..._types.partials.threads import PartialThread
 
 __all__ = ["thread_from_create_request"]
