+++ src/openai_responses/helpers/builders/vector_stores.py
@@ -4,7 +4,7 @@ import httpx
 from openai.types.beta.vector_store import VectorStore
 
 from ._base import _generic_builder
-from ..._routes.vector_stores import VectorStoreCreateRoute
+from ..._routes.beta.vector_stores import VectorStoreCreateRoute
 from ..._types.partials.vector_stores import PartialVectorStore
 
 __all__ = ["vector_store_from_create_request"]
